Added ElasticSearch Maps Reports
