+++ corehq/apps/reports/standard/maps.py
@@ -6,7 +6,7 @@ from django.utils.translation import ugettext_noop
 from casexml.apps.case.models import CommCareCase
 from corehq.apps.reports.api import ReportDataSource
 from corehq.apps.reports.datatables import DataTablesHeader, DataTablesColumn
-from corehq.apps.reports.generic import GenericReportView, GenericTabularReport
+from corehq.apps.reports.generic import GenericReportView, GenericTabularReport, GetParamsMixin
 from corehq.apps.reports.standard import ProjectReport, ProjectReportParametersMixin
 from corehq.apps.reports.standard.cases.basic import CaseListMixin, CaseListReport
 from dimagi.utils.modules import to_function
@@ -252,7 +252,7 @@ class GenericMapReport(ProjectReport, ProjectReportParametersMixin):
         return True
 
 
-class ElasticSearchMapReport(GenericTabularReport, GenericMapReport):
+class ElasticSearchMapReport(GetParamsMixin, GenericTabularReport, GenericMapReport):
 
     report_template_path = "reports/async/maps.html"
     report_partial_path = "reports/async/partials/maps.html"
@@ -271,32 +271,6 @@ class ElasticSearchMapReport(GenericTabularReport, GenericMapReport):
     def total_records(self):
         report = self.get_report()
         return report.total_records
-
-    @property
-    def rows(self):
-        data = self._get_data()
-        columns = self.display_config['table_columns']
-        rows = []
-
-        for feature in data['features']:
-            row = []
-            for column in columns:
-                if column in feature['properties']:
-                    row.append(feature['properties'][column])
-                else:
-                    disp_col = '__disp_' + column
-                    if disp_col in feature['properties']:
-                        row.append(feature['properties'][disp_col])
-            rows.append(row)
-        return rows
-
-    @property
-    def headers(self):
-        columns = self.display_config['table_columns']
-        headers = DataTablesHeader(*[
-            DataTablesColumn(name=name, sortable=False) for name in columns]
-        )
-        return headers
 
     @property
     def json_dict(self):
