+++ corehq/apps/reports/templates/reports/async/maps.html
@@ -1,36 +1,8 @@
-{% extends report_base|default:"reports/async/default.html" %}
+{% extends "reports/async/tabular.html" %}
 {% load hq_shared_tags %}
 {% load report_tags %}
 {% load i18n %}
 
-{% block js %}{{ block.super }}
-    {% if charts %}
-    <link href="{% static 'hqwebapp/js/lib/nvd3/nv.d3.css' %}" rel="stylesheet">
-    <script src="{% static 'hqwebapp/js/lib/nvd3/lib/d3.v2.js' %}"></script>
-    <script src="{% static 'hqwebapp/js/lib/nvd3/nv.d3.min.js' %}"></script>
-    {% endif %}
-{% endblock %}
-
-{% block reportcontent %}
-{% block pretable %}
-    {% if not report.needs_filters %}
-    <h4 class="media-heading">{{ report.report_title }}
-        {% for subtitle in report.report_subtitles %}
-        <br/><small>{{ subtitle }}</small>
-        {% endfor %}
-    </h4>
-    {% endif %}
-    {% if charts %}
-    <div class="row">
-        {% for chart in charts %}
-        <div id='chart_{{ report.slug }}_{{ forloop.counter }}' class="span{{ chart_span }} hide">
-            {% if chart.title %}<h4 style="text-align: center;">{{ chart.title }}</h4>{% endif %}
-            <svg style='height: {{ chart.height }}px'> </svg>
-        </div>
-        {% endfor %}
-    </div>
-    {% endif %}
-{% endblock %}
 {% block reporttable %}
 {% include report_partial %}
 {% if report.needs_filters %}
@@ -98,52 +70,4 @@
         {% endif %}
     </table>
 {% endif %}
-{% endblock %}
-{% block posttable %}{% endblock %}
-{% endblock %}
-
-{% block js-inline %} {{ block.super }}
-    <script type="text/javascript">
-
-        {% if report_table and report_table.datatables %}
-            var reportTables = new HQReportDataTables({
-                dataTableElem: '#report_table_{{ report.slug }}',
-                defaultRows: {{ report_table.default_rows|default:10 }},
-                startAtRowNum: {{ report_table.start_at_row|default:0 }},
-                showAllRowsOption: {{ report_table.show_all_rows|JSON }},
-
-                {% if report_table.headers.render_aoColumns %}aoColumns: {{ report_table.headers.render_aoColumns|JSON }},{% endif %}
-                autoWidth: {{ report_table.headers.auto_width|JSON }},
-                {% if report_table.headers.custom_sort %}customSort: {{ report_table.headers.custom_sort|JSON }},{% endif %}
-
-                {% if report_table.pagination.is_on %}
-                    ajaxSource: '{{ report_table.pagination.source }}',
-                    ajaxParams: {{ report_table.pagination.params|JSON }},
-                {% endif %}
-
-                {% if report_table.left_col.is_fixed %}
-                    fixColumns: true,
-                    fixColsNumLeft: {{ report_table.left_col.fixed.num }},
-                    fixColsWidth: {{ report_table.left_col.fixed.width }},
-                {% endif %}
-            });
-            if (typeof standardHQReport !== 'undefined') {
-                standardHQReport.handleTabularReportCookies(reportTables);
-            }
-            reportTables.render();
-
-        {% endif %}
-
-        $(function() {
-            $('.header-popover').popout({
-                trigger: 'hover',
-                placement: 'bottom'
-            });
-        });
-    </script>
-    {% for chart in charts %}
-        {% with id=forloop.counter|stringformat:"s" slug=report.slug %}
-            {% include chart.template_partial with chart=chart chart_id='chart_'|add:slug|add:'_'|add:id %}
-        {% endwith %}
-    {% endfor %}
 {% endblock %}
