+++ corehq/apps/reports/templates/reports/async/partials/maps.html
@@ -111,11 +111,11 @@ MIN_HEIGHT = 300; //px
   <div id="info" class="control-pane" style="display: none;">
   </div>
 
-  <div id="zoomtofit" class="leaflet-control-layers">
+  <div id="zoomtofit" class="leaflet-control-layers" style="display: none;">
     <div id="zoomtofit-target" class="zoomtofit leaflet-control-layers-toggle" title="Fit all data into view"></div>
   </div>
 
-  <div id="toggletable" class="leaflet-control-layers">
+  <div id="toggletable" class="leaflet-control-layers" style="display: none;">
     <div id="toggletable-target" class="toggletable leaflet-control-layers-toggle" title="Show/hide data table">
       <i class="icon-table"></i>
     </div>
